requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyjwt[crypto]==2.9.0
httpx==0.27.2
azure-cosmos==4.7.0
azure-identity==1.17.1
python-dotenv==1.0.1
